requirements.txt
asgiref
beautifulsoup4
certifi
charset-normalizer
cloudinary
decouple
dj-database-url
Django
django-cloudinary-storage
django-cors-headers
django-filter
djangorestframework
djangorestframework_simplejwt
idna
pillow
psycopg
psycopg-binary
psycopg2-binary
PyJWT
python-decouple
python-dotenv
requests
rest-framework-simplejwt
six
soupsieve
sqlparse
stripe
typing_extensions
urllib3
